Rewrite import_ndex function for speedup Rewritten function and starting to split the tasks.py file into multiple ones
