+++ apps/clustering/tasks/__init__.py
@@ -1,0 +1,1 @@
+from .tasks import run_algorithm
